refactor: rename list node folder
